move: BoundedPool class hierarchy out of ComprehensionExpr and into own directory
